API

Base.:*

Base.:+

Base.hcat

Base.hvcat

Base.propertynames

Base.vcat

Deneb.Data

Deneb.Encoding

Deneb.Facet

Deneb.Mark

Deneb.Params

Deneb.Repeat

Deneb.Transform

Deneb.concat

Deneb.condition

Deneb.config

Deneb.expr

Deneb.field

Deneb.interactive_scales

Deneb.layout

Deneb.param

Deneb.print_theme

Deneb.projection

Deneb.resolve

Deneb.resolve_axis

Deneb.resolve_legend

Deneb.resolve_scale

Deneb.select

Deneb.select_bind_input

Deneb.select_checkbox

Deneb.select_dropdown

Deneb.select_interval

Deneb.select_legend

Deneb.select_point

Deneb.select_radio

Deneb.select_range

Deneb.set_theme!

Deneb.spec

Deneb.title

Deneb.transform_fold

Deneb.transform_window

Deneb.vlspec

Base.:* — Method

spec1::AbstractSpec * spec2::AbstractSpec

Multiplication of two AbstractSpec creates a new AbstractSpec as a composition of the two specifications. For instance, vlspec(mark=:bar) * vlspec(title="chart") will be equivalent to vlspec(mark=:bar, title="chart"). Properties defined in spec2 have precedence over spec1, meaning that if a given property is specified in both, then the result specification will use the property from spec2. If the types of the two specs are different then the result spec will be a VegaLiteSpec.

Base.:+ — Method

spec1::ConstrainedSpec + spec2::ConstrainedSpec

The addition of two specs will produce a new spec with both specs layered. The order matters as spec1 will appear below spec2. If the specs contain common data, encoding or size properties, they will be promoted to the top level specification. Layering layered specification with shared data/encoding/sizes will append the layers ([s1, s2] + [s3, s4] -> [s1, s2, s3, s4]), otherwise a nested layer is created ([s1, s2, [s3, s4]]). Multi-view layout specs (facet, repeat, concat) cannot be layered.

Base.hcat — Method

hcat(A::AbstractSpec...)
[spec1 spec2 spec3 ...]

Horizontal concatenation of specs.

Base.hvcat — Method

hvcat(A::AbstractSpec...)
[spec1 spec2; spec3 spec4 ...]

General (wrappable) concatenation of specs.

Base.propertynames — Method

propertynames(::AbstractSpec)

Return the parent properties of a specification.

Base.vcat — Method

vcat(A::AbstractSpec...)
[spec1; spec2; spec3 ...]

Vertical concatenation of specs.

Deneb.Data — Method

Data(table)
Data(; url, [format], [name])
Data(generator::SymbolOrString; properties...)

Deneb.Encoding — Method

Encoding(x; kw...)
Encoding(x, y)
Encoding(; spec...)

Deneb.Facet — Method

Facet(; row, kw...)
Facet(; column, kw...)
Facet(field; columns=nothing, kw...)

Deneb.Mark — Method

Mark(type; kw...)
Mark(; spec...)

Deneb.Params — Method

Params(; spec...)

Deneb.Repeat — Method

Repeat(; row, kw...)
Repeat(; column, kw...)
Repeat(field; columns=nothing, kw...)
Repeat(Vector{String}; columns=nothing, kw...)

Deneb.Transform — Method

Transform(; spec...)

Deneb.concat — Method

concat(A::AbstractSpec...; columns)

Deneb.condition — Function

condition(param::String, iftrue, iffalse)
condition_test(test::String, iftrue, iffalse)

If iftrue/iffalse isn't a NamedTuple, then it'll be converted as a NamedTuple with name :value.

Deneb.config — Method

config(; properties...)
config(type; properties...)

Deneb.expr — Method

expr(expr)

Convenient function to create an expr spec: {"expr": expr}.

Deneb.field — Method

field(field; kw...)

Shortcut to create an arbitrary encoding channel with a field.

Deneb.interactive_scales — Method

interactive_scales(;bindx=true, bindy=true, shift_on_y=false)

Creates a ParamsSpec that can be composed to other specs to create interactive pan (mouse hold and drag) and zoom (mouse wheel) charts. bindx and bindy specify if the x and y channels are to be bound. If shift_on_y is true, then the shift key must be hold to pan and zoom in the y channel.

Deneb.layout — Method

layout(; align, bounds, center, spacing, columns)

Set layout properties. Needs to be composed with a LayoutSpec (Repeat, Facet, Concat).

Deneb.param — Method

param(param)

Convenient function to create a param spec: {"param": param}.

Deneb.print_theme — Function

print_theme()

Prints the specification of the current theme.

Deneb.projection — Method

projection(type; kw...)

Sets the projection properties.

Deneb.resolve — Method

resolve(type; channels...)
resolve(type, channel, option)

Creates a ResolveSpec. The type indicates the resolution to be defined: scale, axis, or legend.

For scales, resolution can be specified for every channel. For axes, resolutions can be defined for positional channels (x, y, xOffset, yOffset). For legends, resolutions can be defined for non-positional channels (color, opacity, shape, and size).

There are two options to resolve a scale, axis, or legend: shared and independent. Independent scales imply independent axes and legends.

The defaults are documented in Vega-Lite's documentation.

Example

resolve(:scale, color=:independent)

Deneb.resolve_axis — Method

resolve_axis(; channels...)

Alias to resolve(:axis; channels...)

Deneb.resolve_legend — Method

resolve_legend(; channels...)

Alias to resolve(:legend; channels...)

Deneb.resolve_scale — Method

resolve_scale(; channels...)

Alias to resolve(:scale; channels...)

Deneb.select — Method

select(type, name; value, bind, select_options...)

Convenient function to create a ParamsSpec with the following structure:

{
  "name": name,
  "value": value,
  "select": {
    "type": type,
    select_options...
  },
  "bind": bind
}

Deneb.select_bind_input — Method

select_bind_input(type, name; value, select, bind_options...)

Convenient function to create a ParamsSpec with the following structure:

{
  "name": name,
  "value": value,
  "select": select,
  "bind": {
    "input": type,
    bind_options...
  }
}

Deneb.select_checkbox — Method

select_checkbox(type, name; value, select, bind_options...)

Deneb.select_dropdown — Method

select_dropdown(type, name; options, value, select, bind_options...)

options is required

Deneb.select_interval — Method

select_interval()

Deneb.select_legend — Method

select_legend(name; encodings=:color, fields=nothing, bind_options=nothing)

Creates a ParamSpec named name that can be composed to other specs to create selectable legends bound to the given encoding or field. To customize the events that trigger legend interaction, set bind_options with a property that maps to a Vega event stream (e.g. "dblclick"). More info about legend binding: https://vega.github.io/vega-lite/docs/bind.html#legend-binding

Deneb.select_point — Method

select_point()

Deneb.select_radio — Method

select_radio(type, name; options, value, select, bind_options...)

options is required

Deneb.select_range — Method

select_range(type, name; value, select, bind_options...)

Deneb.set_theme! — Method

set_theme!(theme::Symbol)

Sets the current theme to be used. The default Deneb's theme (:default) sets the following global config:

{
    "view": {"continuousWidth": 300, "continuousHeight": 300, "step": 25},
    "mark": {"tooltip": true}
}

The :default_no_tooltip theme sets the default size as above, but disables tooltips. The :empty theme uses Vega-Lite default empty configuration.

Other available themes are any of the Vega themes: :dark, :excel, :fivethirtyeight, :ggplot2, :googlecharts, :latimes, :powerbi, :quartz, :urbaninstitute, :vox (see https://vega.github.io/vega-themes for more info). In this case the :default global config will be used.

set_theme!(config_theme::Symbol, vega_theme::Symbol)

Sets the global config theme (:default, :default_no_tooltip, :empty) and the vega theme (use vega_theme = :default for no vega theme).

set_theme!(config::NamedTuple, [vega_theme::Symbol])

Sets a theme with a user-specified config.

Deneb.spec — Method

spec(s)

Creates an arbitrary Vega-Lite spec.

Deneb.title — Method

title(title)

Deneb.transform_fold — Method

transform_fold(fold::Vector; as=(:key, :value))

Wide to long transformation

Deneb.transform_window — Method

if a sortby field starts with '-' then descending order

Deneb.vlspec — Method

vlspec(s)

Creates a Vega-Lite spec enforcing certain Vega-Lite constrains.